requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
attrs==18.2.0
black==18.9b0
certifi==2018.11.29
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.7
fake-useragent==0.1.11
future==0.17.1
idna==2.8
isort==4.3.4
oauthlib==3.0.1
pycparser==2.19
python-dateutil==2.8.1
python-telegram-bot==11.1.0
python-twitter==3.5
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
soupsieve==1.8
telegram-send==0.20
toml==0.10.0
urllib3==1.24.2
